Move tests for the FCoE module
Move the unit tests for the FCoE module to a new file.
